requirements.txt
flask==2.3.2
openai==0.27.8
python-dotenv==1.0.0
flask-session==0.4.0
Pillow==9.5.0
werkzeug==2.3.7